@@ -1,9 +1,11 @@
 from os import listdir
-from os.path import  basename, dirname, exists, isfile, join
+from os.path import basename, dirname, exists, isfile, join
 from importlib.machinery import SourceFileLoader
 import sys
 
 from abc import ABC, abstractmethod
+
+
 class Derivative(ABC):
     """Derivative - abstract base class and registry for concrete derivatives, which are processors that create new files or accomplish processing tasks based on the contents of a mailbox.
 
@@ -20,12 +22,13 @@ class Derivative(ABC):
 
     def __init_subclass__(cls, **kwargs):
         """Enforce format_name attribute on subclasses, register them"""
-        if not hasattr(cls, 'derivative_name'):
-            raise RuntimeError("Derivative subclass must have `derivative_name` attribute")
+        if not hasattr(cls, "derivative_name"):
+            raise RuntimeError(
+                "Derivative subclass must have `derivative_name` attribute"
+            )
 
         super().__init_subclass__(**kwargs)
         __class__.registry[cls.derivative_name] = cls
-
 
     @abstractmethod
     def __init__(self, email_account, **kwargs):
@@ -34,7 +37,6 @@ class Derivative(ABC):
         If not taking any other steps, the implementation of this can simple be `super()`
         """
         self.account = email_account
-
 
     @abstractmethod
     def do_task_per_account(self):
@@ -46,22 +48,26 @@ class Derivative(ABC):
         """Perform any tasks that should happen once per message"""
         pass
 
+
 def import_derivatives(additional_dirs=None):
     if not additional_dirs:
         additional_dirs = []
 
-    dirs = [join(dirname(__file__), 'derivatives'), *additional_dirs]
+    dirs = [join(dirname(__file__), "derivatives"), *additional_dirs]
 
     for derivatives_dir in dirs:
-        if not exists(derivatives_dir): continue
+        if not exists(derivatives_dir):
+            continue
 
         for filename in listdir(derivatives_dir):
             module = basename(filename)[:-3]
             full_path = join(derivatives_dir, filename)
             # skip if not a normal, non underscored file ending in .py
-            if module.startswith('_') or \
-               not isfile(full_path) or \
-               filename[-3:] != '.py':
+            if (
+                module.startswith("_")
+                or not isfile(full_path)
+                or filename[-3:] != ".py"
+            ):
                 continue
 
             SourceFileLoader(module, full_path).load_module()
