@@ -1,7 +1,7 @@
-#This is Eml derivative
+# This is Eml derivative
 from os.path import join
 import mailbag.helper as helper
-import os,glob
+import os, glob
 import mailbox
 from mailbag.email_account import EmailAccount
 from structlog import get_logger
@@ -14,11 +14,13 @@ from mailbag.derivative import Derivative
 
 
 log = get_logger()
-class EmlDerivative(Derivative):
-    derivative_name = 'eml'
-    derivative_format = 'eml'
 
-    def __init__(self,email_account, **kwargs):
+
+class EmlDerivative(Derivative):
+    derivative_name = "eml"
+    derivative_format = "eml"
+
+    def __init__(self, email_account, **kwargs):
         log.debug("Setup account")
         super()
 
@@ -31,7 +33,6 @@ class EmlDerivative(Derivative):
     def do_task_per_account(self):
         print(self.account.account_data())
 
-
     def do_task_per_message(self, message):
 
         out_dir = os.path.join(self.eml_dir, message.Derivatives_Path)
@@ -39,12 +40,15 @@ class EmlDerivative(Derivative):
 
         # Build msg
         if not message.Message and not message.Headers:
-            log.error("Unable to create EML as no body or headers present for " + str(message.Mailbag_Message_ID))
+            log.error(
+                "Unable to create EML as no body or headers present for "
+                + str(message.Mailbag_Message_ID)
+            )
         else:
             if message.Message:
                 msg = message.Message
             elif message.Headers:
-                msg = MIMEMultipart('mixed')
+                msg = MIMEMultipart("mixed")
                 folder_header = False
                 for key in message.Headers:
                     value = message.Headers[key]
@@ -56,37 +60,54 @@ class EmlDerivative(Derivative):
 
                 # Add message body
                 if message.HTML_Body or message.Text_Body:
-                    alt = MIMEMultipart('alternative')
+                    alt = MIMEMultipart("alternative")
                     if message.Text_Body:
-                        alt.attach(MIMEText(message.Text_Body, 'plain', message.Text_Encoding))
+                        alt.attach(
+                            MIMEText(message.Text_Body, "plain", message.Text_Encoding)
+                        )
                     if message.HTML_Body:
-                        alt = MIMEMultipart('alternative')
-                        alt.attach(MIMEText(message.HTML_Body, 'html', message.HTML_Encoding))
+                        alt = MIMEMultipart("alternative")
+                        alt.attach(
+                            MIMEText(message.HTML_Body, "html", message.HTML_Encoding)
+                        )
                     msg.attach(alt)
                 else:
-                    log.warn("No body present for " + str(message.Mailbag_Message_ID) + ". Created EML without message body.")
+                    log.warn(
+                        "No body present for "
+                        + str(message.Mailbag_Message_ID)
+                        + ". Created EML without message body."
+                    )
 
                 # Attachments
                 for attachment in message.Attachments:
                     mimeType = attachment.MimeType
                     if mimeType is None:
                         mimeType = "text/plain"
-                        log.warn("Mime type not found for the attachment. Set as "+mimeType+".")
-                    mimeType = mimeType.split('/')
+                        log.warn(
+                            "Mime type not found for the attachment. Set as "
+                            + mimeType
+                            + "."
+                        )
+                    mimeType = mimeType.split("/")
                     part = MIMEBase(mimeType[0], mimeType[1])
                     part.set_payload(attachment.File)
                     encoders.encode_base64(part)
-                    part.add_header('Content-Disposition', 'attachment', filename=attachment.Name)
+                    part.add_header(
+                        "Content-Disposition", "attachment", filename=attachment.Name
+                    )
                     msg.attach(part)
             else:
-                log.error("Unable to create EML as no body or headers present for " + str(message.Mailbag_Message_ID))
+                log.error(
+                    "Unable to create EML as no body or headers present for "
+                    + str(message.Mailbag_Message_ID)
+                )
 
             # Write EML to disk
             log.debug("Writing EML to " + str(out_dir))
             if not self.args.dry_run:
                 if not os.path.isdir(out_dir):
                     os.makedirs(out_dir)
-                with open(filename + ".eml",'w') as outfile:
+                with open(filename + ".eml", "w") as outfile:
                     gen = generator.Generator(outfile)
                     gen.flatten(msg)
                     outfile.close()
