@@ -6,8 +6,10 @@ log = get_logger()
 
 # Does nothing currently
 from mailbag.derivative import Derivative
+
+
 class ExampleDerivative(Derivative):
-    derivative_name = 'example'
+    derivative_name = "example"
 
     def __init__(self, email_account, **kwargs):
         log.debug("Setup account")
