@@ -15,7 +15,8 @@ log = get_logger()
 
 class Mbox(EmailAccount):
     """Mbox - This concrete class parses mbox file format"""
-    format_name = 'mbox'
+
+    format_name = "mbox"
 
     def __init__(self, target_account, args, **kwargs):
         log.debug("Parsity parse")
@@ -31,7 +32,7 @@ class Mbox(EmailAccount):
         return account_data
 
     def messages(self):
-        
+
         if os.path.isfile(self.file):
             files = self.file
             parent_dir = os.path.dirname(self.file)
@@ -44,13 +45,15 @@ class Mbox(EmailAccount):
 
             data = mailbox.mbox(filePath)
             for mail in data.itervalues():
-                
+
                 attachments = []
                 errors = {}
                 errors["msg"] = []
                 errors["stack_trace"] = []
                 try:
-                    mailObject = email.message_from_bytes(mail.as_bytes(),policy=email.policy.default)
+                    mailObject = email.message_from_bytes(
+                        mail.as_bytes(), policy=email.policy.default
+                    )
 
                     # Try to parse content
                     try:
@@ -61,9 +64,13 @@ class Mbox(EmailAccount):
                         bodies["text_encoding"] = None
                         if mailObject.is_multipart():
                             for part in mailObject.walk():
-                                bodies, attachments, errors = helper.parse_part(part, bodies, attachments, errors)
+                                bodies, attachments, errors = helper.parse_part(
+                                    part, bodies, attachments, errors
+                                )
                         else:
-                            bodies, attachments, errors = helper.parse_part(part, bodies, attachments, errors)
+                            bodies, attachments, errors = helper.parse_part(
+                                part, bodies, attachments, errors
+                            )
                     except Exception as e:
                         desc = "Error parsing message parts"
                         errors = helper.handle_error(errors, e, desc)
@@ -71,24 +78,29 @@ class Mbox(EmailAccount):
                     # Look for message arrangement
                     try:
                         messagePath = helper.messagePath(mailObject)
-                        unsafePath = os.path.join(os.path.splitext(originalFile)[0], messagePath)
+                        if len(messagePath) > 0:
+                            unsafePath = os.path.join(
+                                os.path.splitext(originalFile)[0], messagePath
+                            )
+                        else:
+                            unsafePath = os.path.splitext(originalFile)[0]
                         derivativesPath = helper.normalizePath(unsafePath)
                     except Exception as e:
                         desc = "Error reading message path from headers"
                         errors = helper.handle_error(errors, e, desc)
-                    
+
                     message = Email(
                         Error=errors["msg"],
-                        Message_ID=mail['Message-ID'].strip(),
+                        Message_ID=mail["Message-ID"].strip(),
                         Original_File=originalFile,
                         Message_Path=messagePath,
                         Derivatives_Path=derivativesPath,
-                        Date=mail['Date'],
-                        From=mail['From'],
-                        To=mail['To'],
-                        Cc=mail['Cc'],
-                        Bcc=mail['Bcc'],
-                        Subject=mail['Subject'],
+                        Date=mail["Date"],
+                        From=mail["From"],
+                        To=mail["To"],
+                        Cc=mail["Cc"],
+                        Bcc=mail["Bcc"],
+                        Subject=mail["Subject"],
                         Content_Type=mailObject.get_content_type(),
                         Headers=mail,
                         HTML_Body=bodies["html_body"],
@@ -97,14 +109,13 @@ class Mbox(EmailAccount):
                         Text_Encoding=bodies["text_encoding"],
                         Message=mailObject,
                         Attachments=attachments,
-                        StackTrace = errors["stack_trace"]
+                        StackTrace=errors["stack_trace"],
                     )
                 except (email.errors.MessageParseError, Exception) as e:
-                    desc = 'Error parsing message'
+                    desc = "Error parsing message"
                     errors = helper.handle_error(errors, e, desc)
                     message = Email(
-                        Error=errors["msg"],
-                        StackTrace=errors["stack_trace"]
+                        Error=errors["msg"], StackTrace=errors["stack_trace"]
                     )
                     log.error(error_msg)
 
@@ -113,4 +124,6 @@ class Mbox(EmailAccount):
             # Make sure the MBOX file is closed
             data.close()
             # Move MBOX to new mailbag directory structure
-            new_path = helper.moveWithDirectoryStructure(self.dry_run,parent_dir,self.mailbag_name,self.format_name,filePath)
+            new_path = helper.moveWithDirectoryStructure(
+                self.dry_run, parent_dir, self.mailbag_name, self.format_name, filePath
+            )
